backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
face-recognition==1.3.0
dlib==19.24.2
numpy==1.24.3
Pillow==10.1.0
cloudinary==1.36.0
supabase==2.0.3
python-dotenv==1.0.0
requests==2.31.0